+++ app_configs/config_base.py
@@ -1,0 +1,10 @@
+class ConfigBase(object):
+
+    def load_before(self, root_widget):
+        pass
+
+    def load_after(self, root_widget):
+        pass
+
+    def load_with_args(self, *args, **kwargs):
+        pass